backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Google Services
google-generativeai==0.3.2
google-auth==2.26.1
google-auth-httplib2==0.2.0

# Image Processing
Pillow==10.2.0
numpy==1.26.3

# Payment Processing
stripe==7.8.0
httpx==0.26.0

# OAuth
authlib==1.3.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1